Split files to one type per file
